# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HighTac Local Platform API
  version: 2.0.0
  summary: Contract for the HighTac LAN management platform.
  description: |
    The Windows backend is the source of truth for stations, tags, products,
    bindings, commands, devices, and audit records. Timestamps are RFC 3339 UTC.

    Admin browser requests use the `hightac_session` cookie. Every mutating
    admin request also requires `X-CSRF-Token`. Approved Android installations
    use bearer identity and never receive MQTT credentials. The canonical retry
    key for retry-safe writes is the `Idempotency-Key` header; keys are scoped to
    the authenticated principal and retained for at least 24 hours.

    Approved Android devices also use bearer-authenticated, allowlisted mobile
    proxies for OpenAI-compatible Responses and JianDaoYun catalog reads. Third-
    party API credentials remain server-side and are never returned to clients.
  contact:
    name: HighTac Engineering
  license:
    name: Proprietary
...
servers:
  - url: http://{host}:{port}/api/v1
    description: HighTac platform on the trusted local network
    variables:
      host:
        default: 192.168.1.105
      port:
        default: "8088"
...
paths:
...
  /device-enrollments:
    post:
      tags: [Device enrollments]
      operationId: createDeviceEnrollment
      summary: Create or resume an Android installation enrollment.
      description: Raw ANDROID_ID and private installation key material must never be sent.
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceEnrollmentCreate"
            example:
              fingerprint_hash: b53b9c2e1d61f3f3d6f470c691e630741e4f2ec95de943dc0c9f71bde7ad1f29
              installation_key_hash: 70d130a4cbf06c7f3a1767f4c1e24661bde4cb21e5d0b2b13bd6e82401c373a2
              manufacturer: Example Mobile
              model: Field-10
              app_version: 2.0.0
      responses:
        "202":
          description: Enrollment is pending approval. The poll secret is returned once.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnrollmentCreated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /device-enrollments/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Device enrollments]
      operationId: getDeviceEnrollment
      summary: Poll enrollment state and receive the approved device token once.
      description: The poll secret is installation-scoped and is replaced by Android bearer identity after approval.
      security:
        - EnrollmentPollSecret: []
      responses:
        "200":
          description: Current enrollment state. `device_token` appears only on the first approved response.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnrollmentStatus"
              examples:
                pending:
                  value:
                    id: 88f22b30-f160-4446-b793-883628620243
                    status: PENDING
                    display_name: null
                    device_id: null
                    device_token: null
                    expires_at: "2026-07-17T08:15:30Z"
                approvedOnce:
                  value:
                    id: 88f22b30-f160-4446-b793-883628620243
                    status: APPROVED
                    display_name: Receiving Phone
                    device_id: ba17796c-c44e-4dc5-9afc-e82af7ef0cf1
                    device_token: htdev_fixture_token_returned_once
                    expires_at: "2026-07-17T08:15:30Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /android-devices:
    get:
      tags: [Android devices]
      operationId: listAndroidDevices
      summary: List pending, approved, and revoked Android installations.
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DeviceStatus"
      responses:
        "200":
          description: Paginated Android installations.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AndroidDevicePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /android-devices/{id}/approve:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      tags: [Android devices]
      operationId: approveAndroidDevice
      summary: Approve a pending Android installation and assign its display name.
      security:
        - AdminSession: []
          CsrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRename"
            example:
              display_name: Receiving Phone
      responses:
        "200":
          description: Approved Android installation. The bearer token is delivered only through enrollment polling.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AndroidDevice"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /android-devices/{id}/revoke:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      tags: [Android devices]
      operationId: revokeAndroidDevice
      summary: Revoke an Android installation and its bearer token.
      security:
        - AdminSession: []
          CsrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmationRequest"
            example:
              confirmation: REVOKE DEVICE
      responses:
        "200":
          description: Revoked Android installation.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AndroidDevice"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /android-devices/{id}/rename:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      tags: [Android devices]
      operationId: renameAndroidDevice
      summary: Set the field-facing name of an Android installation.
      security:
        - AdminSession: []
          CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRename"
      responses:
        "200":
          description: Renamed Android installation.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AndroidDevice"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: hightac_session
      description: Opaque HttpOnly SameSite=Strict administrator session cookie.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required together with AdminSession on every mutating browser request.
    AndroidBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-token
      description: Approved installation token stored in Android Keystore; only its hash is stored server-side.
    EnrollmentPollSecret:
      type: apiKey
      in: header
      name: X-Enrollment-Secret
      description: Short-lived installation-scoped secret returned once when enrollment is created.
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: Number of records per page.
      schema:
        type: integer
        enum: [20, 50, 100]
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Comma-separated allowlisted fields; prefix a field with `-` for descending order.
      schema:
        type: string
        maxLength: 256
        example: -updated_at
    Search:
      name: q
      in: query
      required: false
      description: Trimmed case-insensitive search text.
      schema:
        type: string
        minLength: 1
        maxLength: 256
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID. Reusing a key with the same method, path, actor,
        and normalized body returns the original result. Reusing it with a
        different request returns 409 IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        format: uuid
      example: 2ec5fd45-0e4d-4be4-a6bd-3517118b64df
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
  headers:
    RequestId:
      description: Correlation identifier for logs and support diagnostics.
      schema:
        type: string
        format: uuid
    Location:
      description: Relative API path of the created or accepted resource.
      schema:
        type: string
        example: /api/v1/light-commands/d7d74ed4-44d7-4934-af3a-a80d93aaf63f
  responses:
...
    BadRequest:
      description: Request syntax or confirmation value is invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: BAD_REQUEST
              message: The request could not be processed.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Valid authentication is required.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Forbidden:
      description: The authenticated actor lacks permission, CSRF validation failed, or the Android device is not approved.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: FORBIDDEN
              message: This actor cannot perform the requested operation.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    NotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Conflict:
      description: Current resource state conflicts with the request or an idempotency key was reused for a different request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            activeBinding:
              value:
                error:
                  code: TAG_ALREADY_BOUND
                  message: The tag already has an active binding; explicit rebind is required.
                  details:
                    - field: tag_id
                      code: ACTIVE_BINDING_EXISTS
                      message: Tag AD100000048F is bound to product 1711A-OLD.
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            idempotencyMismatch:
              value:
                error:
                  code: IDEMPOTENCY_KEY_REUSED
                  message: The idempotency key belongs to a different request.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            migrationPreviewMismatch:
              value:
                error:
                  code: MIGRATION_PREVIEW_MISMATCH
                  message: The preview token does not match this device and record set.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            migrationPreviewStale:
              value:
                error:
                  code: MIGRATION_PREVIEW_STALE
                  message: Authoritative binding state changed after the preview.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Gone:
      description: An enrollment, one-time delivery window, or signed media link has expired.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: One or more path, query, header, or body values failed validation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: VALIDATION_ERROR
              message: One or more values are invalid.
              details:
                - field: tag_id
                  code: PATTERN_MISMATCH
                  message: Must match ^AD1[0-9A-F]{9}$.
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          description: Seconds before another attempt.
          schema:
            type: integer
            minimum: 1
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
      example: 2b16ae29-f7ab-4c44-9fc6-a0ca693d9c68
    Timestamp:
      type: string
      format: date-time
      example: "2026-07-16T08:15:30Z"
...
    Sha256:
      type: string
      pattern: ^[0-9a-f]{64}$
      example: 9e107d9d372bb6826bd81d3542a419d6e6299f6ae24c42b2762c9b6f55d25e5f
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required: [code, message, details, request_id]
      properties:
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]{1,63}$
        message:
          type: string
          minLength: 1
          maxLength: 1024
        details:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          $ref: "#/components/schemas/Uuid"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: [string, "null"]
          maxLength: 256
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]{1,63}$
        message:
          type: string
          minLength: 1
          maxLength: 1024
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          enum: [20, 50, 100]
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    ConfirmationRequest:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation:
          type: string
          minLength: 1
          maxLength: 64
...
    EnrollmentStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED]
    DeviceEnrollmentCreate:
      type: object
      additionalProperties: false
      required: [fingerprint_hash, installation_key_hash, manufacturer, model, app_version]
      properties:
        fingerprint_hash:
          $ref: "#/components/schemas/Sha256"
        installation_key_hash:
          $ref: "#/components/schemas/Sha256"
          description: SHA-256 of the random per-install key; the key itself remains in Android Keystore.
        manufacturer:
          type: string
          minLength: 1
          maxLength: 128
        model:
          type: string
          minLength: 1
          maxLength: 128
        app_version:
          type: string
          minLength: 1
          maxLength: 64
    DeviceEnrollmentCreated:
      type: object
      additionalProperties: false
      required: [id, status, poll_secret, expires_at, poll_after_seconds]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        status:
          type: string
          const: PENDING
        poll_secret:
          type: string
          minLength: 32
          maxLength: 256
          writeOnly: true
        expires_at:
          $ref: "#/components/schemas/Timestamp"
        poll_after_seconds:
          type: integer
          minimum: 1
          maximum: 60
          default: 5
    DeviceEnrollmentStatus:
      type: object
      additionalProperties: false
      required: [id, status, display_name, device_id, device_token, expires_at]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        status:
          $ref: "#/components/schemas/EnrollmentStatus"
        display_name:
          type: [string, "null"]
          maxLength: 128
        device_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        device_token:
          type: [string, "null"]
          minLength: 32
          maxLength: 512
          writeOnly: true
          description: Returned once after approval; the server persists only a token hash.
        expires_at:
          $ref: "#/components/schemas/Timestamp"
    DeviceStatus:
      type: string
      enum: [PENDING, APPROVED, REVOKED]
    AndroidDevice:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
        - manufacturer
        - model
        - app_version
        - status
        - online
        - first_seen_at
        - last_seen_at
        - approved_at
        - revoked_at
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        display_name:
          type: [string, "null"]
          maxLength: 128
        manufacturer:
          type: string
          minLength: 1
          maxLength: 128
        model:
          type: string
          minLength: 1
          maxLength: 128
        app_version:
          type: string
          minLength: 1
          maxLength: 64
        status:
          $ref: "#/components/schemas/DeviceStatus"
        online:
          type: boolean
        first_seen_at:
          $ref: "#/components/schemas/Timestamp"
        last_seen_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        approved_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        revoked_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
    AndroidDevicePage:
      type: object
      additionalProperties: false
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AndroidDevice"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    DeviceRename:
      type: object
      additionalProperties: false
      required: [display_name]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 128
...